Update layer's modifiers' names for better readability.
- remove duplicated hide method, where one was accepting only a closure and second only value. Now single method accepts both value or closure.
- update tests for new names
- update Navigate class for new layer's modifiers' names
- move layer's modifiers' to separated file
- add class value testKey to remove code duplication
- change file test name from UIView_SwiftUIKItTests to CALayer_SwiftUIKitTests
- fix indention in test file
