+++ Sources/SwiftUIKit/Navigation/Navigate.swift
@@ -282,7 +282,7 @@ public class Navigate {
                         $0.backgroundColor = style.color
                         $0.clipsToBounds = true
                 }
-                .corner(radius: 8)
+                .layer(cornerRadius: 8)
                 
             }
             .padding(padding)
